[YT-26] 프로젝트 폴더, 파일 정리, 불필요한 코드 제거
컴포넌트들의 위치와 이름을 보다 이해하기 쉽게하고 잘 읽히도록 수정불필요한 코드 제거
