+++ src/views/shared/containers/Header/HeaderContainer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import Header from '../components/Header';
+import Header from '../../components/Header';
 
 const HeaderContainer = () => (
   <Container>
